02-samples/concierge_sales_assistant/requirements.txt
uv
strands-agents
strands-agents-tools
langfuse